--- a/common/repositories/character_expedition_lockouts_repository.h
+++ b/common/repositories/character_expedition_lockouts_repository.h
@@ -75,8 +75,6 @@ public:
 		Database& db, const std::vector<uint32_t>& character_ids,
 		const std::string& expedition_name, const std::string& ordered_event_name)
 	{
-		auto joined_character_ids = fmt::join(character_ids, ",");
-
 		auto results = db.QueryDatabase(fmt::format(SQL(
 			SELECT
 				character_id,
@@ -93,9 +91,9 @@ public:
 				FIELD(character_id, {}),
 				FIELD(event_name, '{}') DESC
 		),
-			joined_character_ids,
+			fmt::join(character_ids, ","),
 			Strings::Escape(expedition_name),
-			joined_character_ids,
+			fmt::join(character_ids, ","),
 			Strings::Escape(ordered_event_name)
 		));