--- a/loginserver/CMakeLists.txt
+++ b/loginserver/CMakeLists.txt
@@ -1,4 +1,4 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 3.2)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.12)
 
 SET(eqlogin_sources
 	account_management.cpp