--- a/world/zonelist.cpp
+++ b/world/zonelist.cpp
@@ -311,19 +311,19 @@ void ZSList::SendZoneStatus(const char* to, int16 admin, WorldTCPConnection* con
 		strcpy(locked, "No");
 	}
 
-	fmt::memory_buffer out;
+	std::vector<char> out;
 
 	if (connection->IsConsole()) {
-		fmt::format_to(out, "World Locked: {}\r\n", locked);
+		fmt::format_to(std::back_inserter(out), "World Locked: {}\r\n", locked);
 	}
 	else {
-		fmt::format_to(out, "World Locked: {}^", locked);
+		fmt::format_to(std::back_inserter(out), "World Locked: {}^", locked);
 	}
 	if (connection->IsConsole()) {
-		fmt::format_to(out, "Zoneservers online:\r\n");
+		fmt::format_to(std::back_inserter(out), "Zoneservers online:\r\n");
 	}
 	else {
-		fmt::format_to(out, "Zoneservers online:^");
+		fmt::format_to(std::back_inserter(out), "Zoneservers online:^");
 	}
 
 	int v = 0, w = 0, x = 0, y = 0, z = 0;
@@ -363,7 +363,7 @@ void ZSList::SendZoneStatus(const char* to, int16 admin, WorldTCPConnection* con
 				zone_data_string[0] = 0;
 			}
 
-			fmt::format_to(out,
+			fmt::format_to(std::back_inserter(out),
 				"#{:<3} :: {} :: {}:{:<5} :: {:2} :: {}:{} :: {} :: ({})",
 				zone_server_data->GetID(),
 				is_static_string,
@@ -377,21 +377,20 @@ void ZSList::SendZoneStatus(const char* to, int16 admin, WorldTCPConnection* con
 				);
 
 			if (out.size() >= 3584) {
-				auto output = fmt::to_string(out);
 				connection->SendEmoteMessageRaw(
 					to,
 					0,
 					AccountStatus::Player,
 					Chat::NPCQuestSay,
-					output.c_str()
+					out.data()
 				);
 				out.clear();
 			}
 			else {
 				if (connection->IsConsole())
-					fmt::format_to(out, "\r\n");
+					fmt::format_to(std::back_inserter(out), "\r\n");
 				else
-					fmt::format_to(out, "^");
+					fmt::format_to(std::back_inserter(out), "^");
 			}
 			x++;
 		}
@@ -400,24 +399,23 @@ void ZSList::SendZoneStatus(const char* to, int16 admin, WorldTCPConnection* con
 				strcpy(zone_data_string, zone_server_data->GetZoneName());
 			else
 				zone_data_string[0] = 0;
-			fmt::format_to(out, "  #{} {}  {}", zone_server_data->GetID(), is_static_string, zone_data_string);
+			fmt::format_to(std::back_inserter(out), "  #{} {}  {}", zone_server_data->GetID(), is_static_string, zone_data_string);
 			if (out.size() >= 3584) {
-				auto output = fmt::to_string(out);
 				connection->SendEmoteMessageRaw(
 					to,
 					0,
 					AccountStatus::Player,
 					Chat::NPCQuestSay,
-					output.c_str()
+					out.data()
 				);
 				out.clear();
 			}
 			else {
 				if (connection->IsConsole()) {
-					fmt::format_to(out, "\r\n");
+					fmt::format_to(std::back_inserter(out), "\r\n");
 				}
 				else {
-					fmt::format_to(out, "^");
+					fmt::format_to(std::back_inserter(out), "^");
 				}
 			}
 			x++;
@@ -427,21 +425,20 @@ void ZSList::SendZoneStatus(const char* to, int16 admin, WorldTCPConnection* con
 	}
 
 	if (connection->IsConsole()) {
-		fmt::format_to(out, "{} servers listed. {} servers online.\r\n", x, y);
+		fmt::format_to(std::back_inserter(out), "{} servers listed. {} servers online.\r\n", x, y);
 	}
 	else {
-		fmt::format_to(out, "{} servers listed. {} servers online.^", x, y);
+		fmt::format_to(std::back_inserter(out), "{} servers listed. {} servers online.^", x, y);
 	}
 
-	fmt::format_to(out, "{} zones are static zones, {} zones are booted zones, {} zones available.", z, w, v);
+	fmt::format_to(std::back_inserter(out), "{} zones are static zones, {} zones are booted zones, {} zones available.", z, w, v);
 
-	auto output = fmt::to_string(out);
 	connection->SendEmoteMessageRaw(
 		to,
 		0,
 		AccountStatus::Player,
 		Chat::NPCQuestSay,
-		output.c_str()
+		out.data()
 	);
 }