--- a/zone/lua_mob.cpp
+++ b/zone/lua_mob.cpp
@@ -3308,7 +3308,7 @@ luabind::scope lua_register_special_abilities() {
 	return luabind::class_<SpecialAbilities>("SpecialAbility")
 
 		.enum_("constants")
-		[
+		[(
 				luabind::value("summon", static_cast<int>(SPECATK_SUMMON)),
 				luabind::value("enrage", static_cast<int>(SPECATK_ENRAGE)),
 				luabind::value("rampage", static_cast<int>(SPECATK_RAMPAGE)),
@@ -3360,7 +3360,7 @@ luabind::scope lua_register_special_abilities() {
 				luabind::value("immune_aggro_npc", static_cast<int>(IMMUNE_AGGRO_NPC)),
 				luabind::value("modify_avoid_damage", static_cast<int>(MODIFY_AVOID_DAMAGE)),
 				luabind::value("immune_open", static_cast<int>(IMMUNE_OPEN))
-		];
+		)];
 }
 
 #endif