--- a/zone/task_manager.cpp
+++ b/zone/task_manager.cpp
@@ -274,7 +274,7 @@ bool TaskManager::SaveClientState(Client *client, ClientTaskState *cts)
 		return false;
 	}
 
-	const char *ERR_MYSQLERROR = "[TASKS]Error in TaskManager::SaveClientState {}";
+	constexpr const char *ERR_MYSQLERROR = "[TASKS]Error in TaskManager::SaveClientState {}";
 
 	int character_id = client->CharacterID();