We're currently on C++17 - this PR enables C++20

Props to KLS for taking FMT from `7.1.3` the rest of the way to `9.1.0`. This PR gives us a full minute back in compilation time (results vary on setup, compiler etc)

See below for compilation benchmark timings

**Changes**

* Minimum Cmake version to `3.12`
* `std::is_pod` is deprecated and swapped out for `std::is_standard_layout` in a class that is not currently in use in the code
* MSVC or Cmake does some strange stuff with preprocessors that are no longer available when enabling CPP20 that required some cmake adjustments.
* Silence many warnings in Windows that largely are noise and get ignored
* Bump FMT library to 9.1.0 from 5.3.0
* Address warning found in Lua bind files `warning: top-level comma expression in array subscript is deprecated` via https://stackoverflow.com/questions/63000052/comma-operator-in-subscript-operator/63000331#63000331

**Features**

Here is a reference of features within C++20 if interested https://en.cppreference.com/w/cpp/20

Specifically, I'm interested in `designated initializers` as it makes code readability of loading data structures much cleaner.

**Before**

```
static struct my_member_t my_global_three[] = {
    { NULL, "one"},
    { NULL, "two"},
    { NULL, "three"},
};
```

**After**

```
static struct my_member_t my_global_three[] = {
    {.type = NULL, .name = "one"},
    {.type = NULL, .name = "two"},
    {.type = NULL, .name = "three"},
};
```

**Linux**

Did some build testing for anyone else's notes

**Full Recompile (no cache, GCC 10.2.1)**

```
# cpp17 with fmt 5.3.0 (3m7.506s)
# cpp20 with fmt 5.3.0 (3m35.234s)
# cpp17 with fmt 7.1.3 (2m56.933s)
# cpp20 with fmt 7.1.3 (3m6s)
# cpp20 with fmt 9.1.0 (2m10s)
# cpp20 with fmt 9.1.0 (2m9.704s)
```

**Partial Recompiles (Core header change) (no cache, GCC 10.2.1)**

```
# cpp20 with fmt 5.3.0 (38s) (core header change)
# cpp20 with fmt 7.1.3 (34s) (core header change)
# cpp20 with fmt 9.1.0 (29s) (core header change)
# cpp20 with fmt 9.1.0 (27s) (core header change)
```

**Windows**

```
========== Build: 22 succeeded, 0 failed, 2 up-to-date, 1 skipped ==========
```